requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
python-dotenv